requirements.txt
python-telegram-bot
python-dotenv
google-cloud-vision
google-cloud-aiplatform
Pillow
cryptography
